backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
torch==2.0.1
transformers==4.35.0
scikit-learn==1.3.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1
Werkzeug==2.3.7
requests==2.31.0
